requirements.txt
python-dotenv
flask
flask-socketio
supabase
gunicorn